Add files via upload
Linear regression is the stepping stone for building any statistical learning model or Machine Learning Model. Although it is somewhat dull in comparison to other vastly popular and complex learning algorithms, Linear regression is still a very useful tool for predicting quantitative responses. We can easily make good inferences on which variables or predictors are highly responsible for driving the response in good faith. The term *linear* is not simply limited to applying only linear terms of variables as the predictors but also we can fit polynomial versions of the variables as predictors to build a more complex model to capture the behavior of data. In this workshop, We will start with the *Simple Linear Regression* where we will make predictions based on a single variable. Later, we will introduce multiple linear regression to accommodate many predictors and interaction terms in a single model.